+++ src/agentic_flights/__init__.py
@@ -1,3 +1,17 @@
+from agentic_flights.agent import (
+    AGENT_CONTRACT,
+    AGENT_REPLY_MAX_BYTES,
+    AgentFailureReply,
+    AgentRuntimeManifest,
+    AgentSearchReply,
+    AgentTripRequest,
+    BaggageIntent,
+    DecisionRow,
+    FlightAgent,
+    agent_runtime,
+    encode_agent_reply,
+    whole_itinerary_confirmed,
+)
 from agentic_flights.api import AgentAPI
 from agentic_flights.batch import BatchExecutor
 from agentic_flights.data_layers import (
@@ -61,6 +75,18 @@ from agentic_flights.strategy_engine import (
 
 __all__ = [
     "AgentAPI",
+    "AGENT_CONTRACT",
+    "AGENT_REPLY_MAX_BYTES",
+    "AgentFailureReply",
+    "AgentRuntimeManifest",
+    "AgentSearchReply",
+    "AgentTripRequest",
+    "BaggageIntent",
+    "DecisionRow",
+    "FlightAgent",
+    "agent_runtime",
+    "encode_agent_reply",
+    "whole_itinerary_confirmed",
     "BatchExecutor",
     "BatchReport",
     "AirportReference",